fix: ProvenCompactList when empty
When the ProvenCompactList is empty (i.e. when the Vec<DataKind> is
empty), some functions would panic due to expecting at least one
element. This commits adds the protections to avoid these panics.

Also, this commit adds protection against the possibility a of
user, seending [Proven]CompactList with 'empty' DataKinds, i.e.
kinds that describe type with no blocks e.g. DataKind::Unsigned(0).
The protection is to ignore them when computing the len of the list
and to filter them out when expanding